Create Gcd_of_arr.cpp
Given an integer array nums, return the greatest common divisor of the smallest number and largest number in nums.

The greatest common divisor of two numbers is the largest positive integer that evenly divides both numbers.

Input 1: nums = [2,5,6,9,10]
Output 1: 2
Explanation 1: The smallest number in nums is 2. The largest number in nums is 10. The greatest common divisor of 2 and 10 is 2.