Create fourDivisors.cpp
Given an integer array nums, return the sum of divisors of the integers in that array that have exactly four divisors. If there is no such integer in the array, return 0.

Input 1: nums = [21,4,7]
Output 1: 32
Explanation 1: 21 has 4 divisors: 1, 3, 7, 21. 4 has 3 divisors and 7 has 2 divisors. The answer is the sum of divisors of 21 only.

Input 2: nums = [21,21]
Output 2: 64