Create redundant_().cpp
Given a string of balanced expressions, find if it contains a redundant parenthesis or not. A set of parenthesis is redundant if the same sub-expression is surrounded by unnecessary or multiple brackets. Return true if redundant, else false.



Input: str = “((a+b))”
Output: true
Explanation: ((a+b)) can reduced to (a+b), this Redundant